make categories component and refactor app.js
