+++ src/App.js
@@ -1,42 +1,10 @@
-import './categories.styles.scss';
+import Category from './components/categories/categories.component';
 import CategoryItem from './components/category/category-item.component';
 
-const App = () => {
-  const categories = [
-    {
-      title: 'Hats',
-      id: 1,
-      imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
-    },
-    {
-      title: 'Jackets',
-      id: 2,
-      imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
-    },
-    {
-      title: 'Sneakers',
-      id: 3,
-      imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
-    },
-    {
-      title: 'Womens',
-      id: 4,
-      imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
-    },
-    {
-      title: 'Mens',
-      id: 5,
-      imageUrl: 'https://i.ibb.co/R70vBrQ/men.png',
-    },
-  ];
-
-  return (
-    <div className="categories-container">
-      {categories.map((category) => (
-        <CategoryItem key={category.id} category={category} />
-      ))}
-    </div>
-  );
-};
+const App = () => (
+  <Category>
+    <CategoryItem key="id" />
+  </Category>
+);
 
 export default App;
