+++ src/components/categories/categories.styles.scss
